tvl_backends/tvl-backends-fffr/requirements.txt
numpy==1.16.2
Pillow==5.2.0
pytest==4.3.0
torch==1.1.0
scikit-build==0.9.0
